code cleanup in kuudra api
