+++ src/main/java/at/hannibal2/skyhanni/features/nether/reputationhelper/dailyquest/quest/KuudraQuest.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.features.nether.reputationhelper.dailyquest.quest
 
-import at.hannibal2.skyhanni.features.nether.KuudraAPI.KuudraTier
+import at.hannibal2.skyhanni.features.nether.kuudra.KuudraTier
 
 class KuudraQuest(val kuudraTier: KuudraTier, state: QuestState) :
     Quest(
